Readable suite phase D: legacy cell, docs, consume migrated sources
