+++ src/tests/streams/readable/legacy-main.js
@@ -1,0 +1,10 @@
+// Copyright (c) 2026 Cloudflare, Inc.
+// Licensed under the Apache 2.0 license found in the LICENSE file or at:
+//     https://opensource.org/licenses/Apache-2.0
+
+// Entry point for the legacy (unflagged) cell.
+
+export {
+  legacyConstructorsThrow,
+  legacyNativeStreamsStillWork,
+} from 'legacy-constructors';
