+++ src/tests/streams/readable/main.js
@@ -146,11 +146,14 @@ export {
 export {
   readAllTextRequestSmall,
   readAllTextRequestBig,
+  readAllTextResponseSmall,
+  readAllTextResponseBig,
   readAllTextFailedPull,
   readAllTextFailedStart,
   readAllTextControllerError,
   readAllTextLargeBody,
   bodyConsumptionNormalizesBufferSourceChunks,
+  bodyConsumptionPinsResizableArrayBufferExtent,
   cloneWithStreamBody,
   cancelBodyThenConsume,
   fetchBodyRoundtrip,
