+++ src/workerd/api/streams/streams-test.js
@@ -273,150 +273,6 @@ export const inspect = {
 // Before the fix, this would crash with:
 // "expected state.template tryGet<Ready>() != nullptr; The consumer is either closed or errored."
 //
-// This test simulates the production scenario where:
-// 1. A TransformStream's readable is tee'd
-// 2. The transform function synchronously cancels one of the tee branches
-// 3. When enqueue is called, the push loop tries to push to the cancelled consumer
-export const transformTeeReentrancySynchronousCancel = {
-  async test() {
-    let reader2;
-    let cancelledBranch2 = false;
-
-    // Create a TransformStream whose transform function cancels branch2
-    const ts = new TransformStream({
-      transform(chunk, controller) {
-        // First time through, cancel branch2 BEFORE enqueueing
-        // This simulates the production scenario where user code in the
-        // transform function affects another consumer
-        if (!cancelledBranch2 && reader2) {
-          reader2.cancel('cancelled synchronously in transform');
-          cancelledBranch2 = true;
-        }
-        controller.enqueue(chunk);
-      },
-    });
-
-    const writer = ts.writable.getWriter();
-    const [branch1, branch2] = ts.readable.tee();
-    const reader1 = branch1.getReader();
-    reader2 = branch2.getReader();
-
-    // Start pending reads on both branches
-    const read1Promise = reader1.read();
-    const read2Promise = reader2.read();
-
-    // Write to the transform - this triggers the transform function which:
-    // 1. Cancels branch2 (closing/erroring its consumer)
-    // 2. Calls controller.enqueue() which pushes to all consumers
-    // Before the fix, step 2 would crash when trying to push to cancelled branch2
-    await writer.write('test data');
-
-    // Verify branch1 got the data
-    const result1 = await read1Promise;
-    assert.strictEqual(result1.value, 'test data');
-
-    // branch2 was cancelled, so its read should complete (done or with data before cancel)
-    const result2 = await read2Promise;
-    assert.ok(result2 !== undefined);
-
-    await writer.close();
-    await reader1.cancel();
-  },
-};
-
-// Test with TransformStream to match the production stack trace more closely.
-// The production bug occurred during: TransformStream → enqueue → QueueImpl::push → consumer iteration
-export const transformStreamTeeReentrancy = {
-  async test() {
-    const { readable, writable } = new TransformStream();
-    const writer = writable.getWriter();
-
-    const [branch1, branch2] = readable.tee();
-    const reader1 = branch1.getReader();
-    const reader2 = branch2.getReader();
-
-    // Start pending reads on both branches
-    const read1Promise = reader1.read();
-    const read2Promise = reader2.read();
-
-    // When read1 resolves, cancel branch2
-    // This simulates the production scenario where a .then() handler
-    // attached to the read promise cancels another branch
-    read1Promise.then(() => {
-      reader2.cancel('cancelled during transform');
-    });
-
-    // Write through the transform - this triggers enqueue on the readable side.
-    // Before the fix, this would crash when the push loop tried to push to
-    // the now-cancelled branch2 consumer.
-    await writer.write('transform data');
-
-    // Verify read1 succeeded
-    const result1 = await read1Promise;
-    assert.strictEqual(result1.value, 'transform data');
-
-    // read2 may have received data or be done - either is fine
-    // The important thing is no crash occurred
-    const result2 = await read2Promise;
-    assert.ok(result2 !== undefined);
-
-    await writer.close();
-    await reader1.cancel();
-  },
-};
-
-// Test that multiple writes through a tee'd ReadableStream work correctly
-// even when one branch is cancelled mid-stream
-export const teeWithCancelMidStream = {
-  async test() {
-    let controller;
-    const stream = new ReadableStream({
-      start(c) {
-        controller = c;
-      },
-    });
-
-    const [branch1, branch2] = stream.tee();
-    const reader1 = branch1.getReader();
-    const reader2 = branch2.getReader();
-
-    // Start reads on both branches
-    let read1Promise = reader1.read();
-    let read2Promise = reader2.read();
-
-    // Enqueue first chunk - both branches should get it
-    controller.enqueue('chunk1');
-    const r1a = await read1Promise;
-    const r2a = await read2Promise;
-    assert.strictEqual(r1a.value, 'chunk1');
-    assert.strictEqual(r2a.value, 'chunk1');
-
-    // Now cancel branch2
-    await reader2.cancel('done with branch2');
-
-    // Start another read on branch1
-    read1Promise = reader1.read();
-
-    // Enqueue second chunk - only branch1 should get it
-    // This should not crash even though branch2's consumer is now closed
-    controller.enqueue('chunk2');
-    const r1b = await read1Promise;
-    assert.strictEqual(r1b.value, 'chunk2');
-
-    // Start another read and enqueue third chunk to confirm continued operation
-    read1Promise = reader1.read();
-    controller.enqueue('chunk3');
-    const r1c = await read1Promise;
-    assert.strictEqual(r1c.value, 'chunk3');
-
-    // Close and verify
-    read1Promise = reader1.read();
-    controller.close();
-    const r1d = await read1Promise;
-    assert.strictEqual(r1d.done, true);
-  },
-};
-
 // ============================================================================
 
 export const testCancelPipethrough = {
@@ -488,30 +344,5 @@ export const testCancelPipethrough2 = {
     const result = await reader2.read();
     assert.ok(result.done);
     assert.strictEqual(result.value, undefined);
-  },
-};
-
-export const ResponseTextLargeBody = {
-  async test() {
-    const targetSize = 2147483648 + 1024 * 1024;
-    let sent = 0;
-    const chunkSize = 64 * 1024 * 1024;
-
-    const stream = new ReadableStream({
-      pull(controller) {
-        if (sent >= targetSize) {
-          controller.close();
-          return;
-        }
-        const size = Math.min(chunkSize, targetSize - sent);
-        controller.enqueue(new Uint8Array(size).fill(0x41));
-        sent += size;
-      },
-    });
-
-    await assert.rejects(
-      new Response(stream).text(),
-      (e) => e instanceof RangeError
-    );
   },
 };
