--- src/workerd/api/tests/streams-async-iterator-test.wd-test
@@ -1,14 +1,0 @@
-using Workerd = import "/workerd/workerd.capnp";
-
-const unitTests :Workerd.Config = (
-  services = [
-    ( name = "streams-async-iterator-test",
-      worker = (
-        modules = [
-          (name = "worker", esModule = embed "streams-async-iterator-test.js")
-        ],
-        compatibilityFlags = ["nodejs_compat", "streams_enable_constructors", "transformstream_enable_standard_constructor"],
-      )
-    ),
-  ],
-);
