+++ src/workerd/api/tests/streams-test.js
@@ -16,24 +16,6 @@ export const globalScopeWritableStream = {
   },
 };
 
-export const rs = {
-  async test(ctrl, env) {
-    const resp = await env.subrequest.fetch('http://example.org', {
-      method: 'POST',
-      body: new ReadableStream({
-        expectedLength: 10,
-        start(c) {
-          c.enqueue(enc.encode('hellohello'));
-          c.close();
-        },
-      }),
-    });
-    for await (const _ of resp.body) {
-      // intentionally empty
-    }
-  },
-};
-
 export const ts = {
   async test(ctrl, env) {
     const { readable, writable } = new TransformStream({
@@ -46,27 +28,6 @@ export const ts = {
       method: 'POST',
       body: readable,
     });
-    for await (const _ of resp.body) {
-      // intentionally empty
-    }
-  },
-};
-
-// Regression test for https://github.com/cloudflare/workerd/issues/5113
-export const rsRequest = {
-  async test(ctrl, env) {
-    const resp = await env.subrequest.fetch(
-      new Request('http://example.org', {
-        method: 'POST',
-        body: new ReadableStream({
-          expectedLength: 10,
-          start(c) {
-            c.enqueue(enc.encode('hellohello'));
-            c.close();
-          },
-        }),
-      })
-    );
     for await (const _ of resp.body) {
       // intentionally empty
     }
@@ -128,25 +89,6 @@ export const byobMin = {
   },
 };
 
-export const cancelReadsOnReleaseLock = {
-  async test() {
-    const rs = new ReadableStream();
-    const reader = rs.getReader();
-    const read = reader.read();
-
-    const result = await Promise.allSettled([read, reader.releaseLock()]);
-    strictEqual(result[0].status, 'rejected');
-    strictEqual(
-      result[0].reason.message,
-      'This ReadableStream reader has been released.'
-    );
-    strictEqual(result[1].status, 'fulfilled');
-
-    // Make sure we can still get another reader
-    const _reader2 = rs.getReader();
-  },
-};
-
 export const cancelWriteOnReleaseLock = {
   async test() {
     const ws = new WritableStream({
@@ -173,142 +115,6 @@ export const cancelWriteOnReleaseLock = {
 
     // Make sure we can still get another writer
     const _writer2 = ws.getWriter();
-  },
-};
-
-export const readAllTextRequestSmall = {
-  async test() {
-    const rs = new ReadableStream({
-      pull(c) {
-        c.enqueue(enc.encode('hello '));
-        c.enqueue(enc.encode('world!'));
-        c.close();
-      },
-    });
-    const request = new Request('http://example.org', {
-      method: 'POST',
-      body: rs,
-    });
-    const text = await request.text();
-    strictEqual(text, 'hello world!');
-  },
-};
-
-export const readAllTextResponseSmall = {
-  async test() {
-    const rs = new ReadableStream({
-      pull(c) {
-        c.enqueue(enc.encode('hello '));
-        c.enqueue(enc.encode('world!'));
-        c.close();
-      },
-    });
-    const response = new Response(rs);
-    const text = await response.text();
-    strictEqual(text, 'hello world!');
-  },
-};
-
-export const readAllTextRequestBig = {
-  async test() {
-    const chunks = [
-      'a'.repeat(4097),
-      'b'.repeat(4097 * 2),
-      'c'.repeat(4097 * 4),
-    ];
-    let check = '';
-    const enc = new TextEncoder();
-
-    const rs = new ReadableStream({
-      pull(c) {
-        if (chunks.length === 0) {
-          c.close();
-          return;
-        }
-        const chunk = chunks.shift();
-        check += chunk;
-        c.enqueue(enc.encode(chunk));
-      },
-    });
-    const request = new Request('http://example.org', {
-      method: 'POST',
-      body: rs,
-    });
-    const text = await request.text();
-    strictEqual(text.length, check.length);
-    strictEqual(text, check);
-  },
-};
-
-export const readAllTextResponseBig = {
-  async test() {
-    const chunks = [
-      'a'.repeat(4097),
-      'b'.repeat(4097 * 2),
-      'c'.repeat(4097 * 4),
-    ];
-    let check = '';
-    const enc = new TextEncoder();
-
-    const rs = new ReadableStream({
-      async pull(c) {
-        await scheduler.wait(10);
-        if (chunks.length === 0) {
-          c.close();
-          return;
-        }
-        const chunk = chunks.shift();
-        check += chunk;
-        c.enqueue(enc.encode(chunk));
-      },
-    });
-    const response = new Response(rs);
-    const promise = response.text();
-    const text = await promise;
-    strictEqual(text.length, check.length);
-    strictEqual(text, check);
-  },
-};
-
-export const readAllTextFailedPull = {
-  async test() {
-    const rs = new ReadableStream({
-      async pull(c) {
-        await scheduler.wait(10);
-        throw new Error('boom');
-      },
-    });
-    const response = new Response(rs);
-    await rejects(response.text(), { message: 'boom' });
-  },
-};
-
-export const readAllTextFailedStart = {
-  async test() {
-    const rs = new ReadableStream({
-      async start(c) {
-        await scheduler.wait(10);
-        throw new Error('boom');
-      },
-    });
-    const response = new Response(rs);
-    await rejects(response.text(), { message: 'boom' });
-  },
-};
-
-export const readAllTextFailed = {
-  async test() {
-    const rs = new ReadableStream({
-      async start(c) {
-        await scheduler.wait(10);
-        c.error(new Error('boom'));
-      },
-    });
-    const response = new Response(rs);
-    ok(!rs.locked);
-    const promise = response.text();
-    ok(rs.locked);
-    await rejects(promise, { message: 'boom' });
   },
 };
 
@@ -364,272 +170,10 @@ export const writableStreamGcTraceFinishes = {
   },
 };
 
-export const readableStreamFromAsyncGenerator = {
-  async test() {
-    async function* gen() {
-      await scheduler.wait(10);
-      yield 'hello';
-      await scheduler.wait(10);
-      yield 'world';
-    }
-    const rs = ReadableStream.from(gen());
-    const chunks = [];
-    for await (const chunk of rs) {
-      chunks.push(chunk);
-    }
-    deepStrictEqual(chunks, ['hello', 'world']);
-  },
-};
-
-export const readableStreamFromSyncGenerator = {
-  async test() {
-    const rs = ReadableStream.from(['hello', 'world']);
-    const chunks = [];
-    for await (const chunk of rs) {
-      chunks.push(chunk);
-    }
-    deepStrictEqual(chunks, ['hello', 'world']);
-  },
-};
-
-export const readableStreamFromSyncGenerator2 = {
-  async test() {
-    function* gen() {
-      yield 'hello';
-      yield 'world';
-    }
-    const rs = ReadableStream.from(gen());
-    const chunks = [];
-    for await (const chunk of rs) {
-      chunks.push(chunk);
-    }
-    deepStrictEqual(chunks, ['hello', 'world']);
-  },
-};
-
-export const readableStreamFromAsyncCanceled = {
-  async test() {
-    async function* gen() {
-      let count = 0;
-      try {
-        count++;
-        yield 'hello';
-        count++;
-        yield 'world';
-      } finally {
-        strictEqual(count, 1);
-      }
-    }
-    const rs = ReadableStream.from(gen());
-    const chunks = [];
-    for await (const chunk of rs) {
-      chunks.push(chunk);
-      return;
-    }
-    deepStrictEqual(chunks, ['hello']);
-  },
-};
-
-export const readableStreamFromThrowingAsyncGen = {
-  async test() {
-    async function* gen() {
-      yield 'hello';
-      throw new Error('boom');
-    }
-    const rs = ReadableStream.from(gen());
-    const chunks = [];
-    async function consumeStream() {
-      for await (const chunk of rs) {
-        chunks.push(chunk);
-      }
-    }
-    await rejects(consumeStream, { message: 'boom' });
-    deepStrictEqual(chunks, ['hello']);
-  },
-};
-
-export const readableStreamFromNoopAsyncGen = {
-  async test() {
-    async function* gen() {}
-    const rs = ReadableStream.from(gen());
-    const chunks = [];
-    for await (const chunk of rs) {
-      chunks.push(chunk);
-    }
-    deepStrictEqual(chunks, []);
-  },
-};
-
-// Tests for ReadableStream.from() cancel behavior per WPT spec
-export const readableStreamFromCancelRejectsWhenReturnRejects = {
-  async test() {
-    const rejectError = new Error('return error');
-    const iterable = {
-      async next() {
-        return { value: undefined, done: true };
-      },
-      async return() {
-        throw rejectError;
-      },
-      [Symbol.asyncIterator]() {
-        return this;
-      },
-    };
-
-    const rs = ReadableStream.from(iterable);
-    const reader = rs.getReader();
-
-    await rejects(reader.cancel(), rejectError);
-  },
-};
-
-export const readableStreamFromCancelRejectsWhenReturnThrows = {
-  async test() {
-    const throwError = new Error('return throws');
-    const iterable = {
-      async next() {
-        return { value: undefined, done: true };
-      },
-      return() {
-        throw throwError;
-      },
-      [Symbol.asyncIterator]() {
-        return this;
-      },
-    };
-
-    const rs = ReadableStream.from(iterable);
-    const reader = rs.getReader();
-
-    await rejects(reader.cancel(), (err) => err === throwError);
-  },
-};
-
-export const readableStreamFromCancelRejectsWhenReturnNotMethod = {
-  async test() {
-    const iterable = {
-      async next() {
-        return { value: undefined, done: true };
-      },
-      return: 42, // exists but not callable
-      [Symbol.asyncIterator]() {
-        return this;
-      },
-    };
-
-    const rs = ReadableStream.from(iterable);
-    const reader = rs.getReader();
-
-    await rejects(reader.cancel(), {
-      name: 'TypeError',
-      message: /return/,
-    });
-  },
-};
-
-export const readableStreamFromCancelRejectsWhenReturnNonObject = {
-  async test() {
-    const iterable = {
-      async next() {
-        return { value: undefined, done: true };
-      },
-      async return() {
-        return 42; // fulfills with non-object
-      },
-      [Symbol.asyncIterator]() {
-        return this;
-      },
-    };
-
-    const rs = ReadableStream.from(iterable);
-    const reader = rs.getReader();
-
-    await rejects(reader.cancel(), {
-      name: 'TypeError',
-    });
-  },
-};
-
-export const readableStreamFromCancelResolvesWhenReturnMissing = {
-  async test() {
-    const iterable = {
-      async next() {
-        return { value: undefined, done: true };
-      },
-      // no return method
-      [Symbol.asyncIterator]() {
-        return this;
-      },
-    };
-
-    const rs = ReadableStream.from(iterable);
-    const reader = rs.getReader();
-
-    // Should resolve without error when return() is missing
-    await Promise.all([reader.cancel(), reader.closed]);
-  },
-};
-
 // TODO(streams-tests): the tests below exercise internal body streams
 // (`new Response(...).body`) and standard byte streams, not identity
 // streams. They live here until the readable/body suite restructuring
 // gives them a permanent home.
-
-// Test that canceling a stream rejects body consume function
-export const cancelStreamRejectsBodyConsume = {
-  async test() {
-    const response = new Response('foo bar');
-    const stream = response.body;
-
-    stream.cancel(new Error('a good reason'));
-
-    await rejects(response.text(), TypeError);
-  },
-};
-
-// Test that canceling a reader resolves closed promise
-export const cancelReaderResolvesClosedPromise = {
-  async test() {
-    const response = new Response('foo bar');
-    const stream = response.body;
-    const reader = stream.getReader();
-
-    reader.cancel();
-    const closed = await reader.closed;
-    strictEqual(typeof closed, 'undefined');
-    reader.releaseLock();
-
-    await rejects(response.text(), TypeError);
-  },
-};
-
-// Test that getReader with bad mode throws
-export const getReaderBadModeThrows = {
-  test() {
-    const response = new Response('foo bar');
-    const stream = response.body;
-
-    throws(() => stream.getReader({ mode: 'nope' }), TypeError);
-  },
-};
-
-// Test that stream is locked after getReader() called
-export const streamLockedAfterGetReader = {
-  test() {
-    const response = new Response('foo bar');
-    const stream = response.body;
-
-    const reader = stream.getReader();
-
-    ok(stream.locked);
-
-    throws(() => stream.getReader(), TypeError);
-
-    reader.releaseLock();
-    ok(!stream.locked);
-    reader.releaseLock(); // Second time should be a no-op
-  },
-};
 
 // Test BYOB reader constraints
 export const byobReaderConstraints = {
