+++ src/workerd/api/tests/ts-webstreams-test.js
@@ -482,131 +482,11 @@ export const nativeBackedStreamIntoFetchBody = {
   },
 };
 
-// Body consumption accepts any BufferSource chunk shape, matching the C++ pump:
-// ArrayBuffer, DataView, and non-Uint8Array views contribute their bytes, detached
-// buffers AND detached views are empty (including DataViews, whose byteLength
-// getter throws for detached buffers), and a non-BufferSource chunk fails with
-// the C++ error text.
-export const bodyConsumptionNormalizesBufferSourceChunks = {
-  async test() {
-    const enc = new TextEncoder();
-    const bodyOf = (chunks) =>
-      new ReadableStream({
-        start(c) {
-          for (const chunk of chunks) c.enqueue(chunk);
-          c.close();
-        },
-      });
-    strictEqual(
-      await new Response(bodyOf([enc.encode('ab').buffer])).text(),
-      'ab'
-    );
-    strictEqual(
-      await new Response(
-        bodyOf([new DataView(enc.encode('_cd').buffer, 1)])
-      ).text(),
-      'cd'
-    );
-    strictEqual(
-      (await new Response(bodyOf([new Float64Array(1)])).arrayBuffer())
-        .byteLength,
-      8
-    );
-    const detached = new ArrayBuffer(4);
-    detached.transfer();
-    strictEqual(
-      await new Response(bodyOf([detached, enc.encode('x')])).text(),
-      'x'
-    );
-    const detachedViewBuffer = new ArrayBuffer(4);
-    const detachedView = new DataView(detachedViewBuffer);
-    detachedViewBuffer.transfer();
-    strictEqual(
-      await new Response(bodyOf([detachedView, enc.encode('y')])).text(),
-      'y'
-    );
-    await rejects(new Response(bodyOf(['not bytes'])).text(), {
-      name: 'TypeError',
-      message: 'This ReadableStream did not return bytes.',
-    });
-  },
-};
-
-// A resizable ArrayBuffer chunk is captured at its drained extent: user code
-// growing the buffer after the chunk is collected (but before the parts are
-// assembled) must not desync the copy from the accounted length — a bare
-// length-tracking view over the chunk would make assembly throw RangeError.
-export const bodyConsumptionPinsResizableArrayBufferExtent = {
-  async test() {
-    const rab = new ArrayBuffer(2, { maxByteLength: 8 });
-    new Uint8Array(rab).set([120, 121]); // 'xy'
-    let ctrl;
-    const rs = new ReadableStream({
-      start(c) {
-        ctrl = c;
-      },
-    });
-    const bodyPromise = new Response(rs).arrayBuffer();
-    ctrl.enqueue(rab);
-    // A macrotask boundary: the drain loop collects the chunk (pinning its
-    // extent) before the resize below, and only then sees end-of-stream.
-    await scheduler.wait(0);
-    rab.resize(8);
-    ctrl.close();
-    const buf = await bodyPromise;
-    strictEqual(buf.byteLength, 2);
-    strictEqual(new TextDecoder().decode(buf), 'xy');
-  },
-};
-
-// Body preconditions still apply after unwrap: a disturbed stream is rejected by the
-// Body constructor itself (unwrap deliberately performs no such checks). This throw did
-// NOT happen before the unwrap arm landed (the async-iterable fallback wrapped the
-// stream in a fresh, undisturbed one), so this is a legacy-parity regression test.
-export const disturbedTsStreamIntoResponse = {
-  async test() {
-    const rs = new ReadableStream({
-      start(c) {
-        c.enqueue(new TextEncoder().encode('x'));
-        c.close();
-      },
-    });
-    const reader = rs.getReader();
-    await reader.read();
-    reader.releaseLock();
-    throws(() => new Response(rs), {
-      name: 'TypeError',
-      message: /disturbed/,
-    });
-  },
-};
-
 // Non-stream objects keep their non-stream Body semantics: the brand check rejects them,
 // the OneOf falls through, and a plain object stringifies per spec.
 export const plainObjectBodyStillStringifies = {
   async test() {
     strictEqual(await new Response({}).text(), '[object Object]');
-  },
-};
-
-// Response.clone() with a TypeScript-backed (queued) body: clone happens at the C++
-// layer (Body::clone -> JsReadableStream::tee's TS arm -> the TS tee machinery), and
-// both bodies read the full content independently.
-export const tsStreamBodyClone = {
-  async test() {
-    const encoder = new TextEncoder();
-    const rs = new ReadableStream({
-      start(c) {
-        c.enqueue(encoder.encode('clone'));
-        c.enqueue(encoder.encode(' me'));
-        c.close();
-      },
-    });
-    const response = new Response(rs);
-    const clone = response.clone();
-    const [a, b] = await Promise.all([response.text(), clone.text()]);
-    strictEqual(a, 'clone me');
-    strictEqual(b, 'clone me');
   },
 };
